# swagger.yml
basePath: /
consumes:
- application/json
- application/xml
host: 127.0.0.1:6420
info:
  contact:
    email: skycoin.doe@example.com
    name: Skycoin project
    url: http://127.0.0.1:6420
  description: Skycoin is a next-generation cryptocurrency.
  title: Skycoin REST API.
  version: 0.25.1
paths:
...
  /api/v1/transaction:
    get:
      description: Returns a transaction identified by its txid hash with just id
      operationId: transaction
      parameters:
      - description: transaction hash
        in: query
        name: txid
        required: true
        type: string
      - description: return as a raw encoded transaction.
        in: query
        name: encoded
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Returns a transaction identified by its txid hash.
          schema:
            properties:
              status:
                properties:
                  block_seq:
                    description: If confirmed, the sequence of the block in which
                      the transaction was executed
                    format: int64
                    type: integer
                  confirmed:
                    type: boolean
                  height:
                    description: If confirmed, how many blocks deep in the chain it
                      is. Will be at least 1 if confirmed
                    format: int64
                    type: integer
                  unconfirmed:
                    type: boolean
                type: object
              time:
                format: int64
                type: integer
              txn:
                description: TransactionVerbose has readable transaction data. It
                  adds TransactionStatus to a BlockTransactionVerbose
                properties:
                  inner_hash:
                    type: string
                  inputs:
                    items:
                      type: string
                    type: array
                  length:
                    format: int32
                    type: integer
                  outputs:
                    items:
                      properties:
                        coins:
                          type: string
                        dst:
                          type: string
                        hours:
                          format: int64
                          type: integer
                        uxid:
                          type: string
                    type: array
                  sigs:
                    items:
                      type: string
                    type: array
                  timestamp:
                    format: int64
                    type: integer
                  txid:
                    type: string
                  type:
                    format: int32
                    type: integer
                type: object
        default:
          $ref: '#/responses/genericError'
...
  /api/v1/wallet/transaction:
    post:
      description: Creates a signed transaction
      operationId: walletTransaction
      parameters:
      - in: body
        name: body
        required: true
        schema:
          properties:
            change_address:
              type: string
            hours_selection:
              properties:
                mode:
                  type: string
                share_factor:
                  type: string
                type:
                  type: string
              type: object
            ignore_unconfirmed:
              type: boolean
            to:
              items:
                properties:
                  address:
                    type: string
                  coins:
                    format: int64
                    type: integer
                  hours:
                    format: int64
                    type: integer
              type: array
            wallet:
              properties:
                addresses:
                  items:
                    type: string
                  type: array
                id:
                  type: string
                password:
                  type: string
                unspents:
                  items:
                    type: string
                  type: array
              type: object
      produces:
      - application/json
      responses:
        "200":
          description: Returns blocks between a start and end point.
          schema:
            properties:
              encoded_transaction:
                type: string
              transaction:
                properties:
                  fee:
                    type: string
                  inner_hash:
                    type: string
                  inputs:
                    items:
                      properties:
                        address:
                          type: string
                        block:
                          format: int64
                          type: integer
                        calculated_hours:
                          type: string
                        coins:
                          type: string
                        hours:
                          type: string
                        timestamp:
                          format: int64
                          type: integer
                        txid:
                          type: string
                        uxid:
                          type: string
                    type: array
                  length:
                    format: int64
                    type: integer
                  outputs:
                    items:
                      properties:
                        address:
                          type: string
                        coins:
                          type: string
                        hours:
                          type: string
                        uxid:
                          type: string
                    type: array
                  sigs:
                    items:
                      type: string
                    type: array
                  txid:
                    type: string
                  type:
                    format: int32
                    type: integer
                type: object
        default:
          $ref: '#/responses/genericError'
      security:
      - csrfAuth: []
...
produces:
- application/json
- application/xml
responses:
...
  genericError:
    description: |-
      A GenericError is the default error message that is generated.
      For certain status codes there are more appropriate error structures.
    schema:
      properties:
        code:
          format: int32
          type: integer
          x-go-name: Code
        message:
          type: string
          x-go-name: Message
      type: object
schemes:
- http
- https
securityDefinitions:
  csrfAuth:
    in: header
    name: X-CSRF-TOKEN
    type: apiKey
swagger: "2.0"